pyproject.toml
[project]
name = "pet_video_editing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clip",
    "ftfy>=6.3.1",
    "moviepy>=2.1.2",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=10.4.0",
    "regex>=2024.11.6",
    "tables>=3.10.2",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
